fix(proto): the size of the repo should be int64, since that is the same type used for the manifest/config/index/digest sizes it sums up.
Using int32 may result in negative size values when returned by the graphql API

+++ pkg/meta/convert/convert.go
@@ -445,7 +445,7 @@ func GetRepoMeta(protoRepoMeta *proto_go.RepoMeta) mTypes.RepoMeta {
 		Name:             protoRepoMeta.Name,
 		Tags:             GetTags(protoRepoMeta.Tags),
 		Rank:             int(protoRepoMeta.Rank),
-		Size:             int64(protoRepoMeta.Size),
+		Size:             protoRepoMeta.Size,
 		Platforms:        GetPlatforms(protoRepoMeta.Platforms),
 		Vendors:          protoRepoMeta.Vendors,
 		IsStarred:        protoRepoMeta.IsStarred,
